refactor get localstate
