+++ examples/example-playground.html
@@ -16,7 +16,7 @@
                 class="{{ fps > 60 ? 'text-green-600' : 'text-yellow-600' }} fixed top-4 right-4"
             >
                 FPS: {{ fps }}
-            </div>
+            </div> -->
             <template id="my-template">
                 <div>Hello {{ dataName }}! {{ dataCount }}</div>
             </template>
@@ -28,7 +28,7 @@
                     </div>
                     <div class="pl-2">{{ children }}</div>
                 </div>
-            </template> -->
+            </template>
             <template id="x-child2">
                 <div class="flex">
                     <div class="pr-2">{{ children }}.</div>
@@ -38,7 +38,7 @@
                     </div>
                 </div>
             </template>
-            <!-- <template id="x-parent">
+            <template id="x-parent">
                 <div class="p-2">
                     {{ names.map( (n, i) => `
                     <x-child data-count="{{ dataParent }}" data-child="${i}"
@@ -46,19 +46,19 @@
                     </x-child>
                     `)}}
                 </div>
-            </template> -->
-            <!-- <my-template
+            </template>
+            <my-template
                 data-name="World"
                 data-count="{{ count }}"
-            ></my-template> -->
-            <!-- <div>Count: {{ count }}</div>
+            ></my-template>
+            <div>Count: {{ count }}</div>
             <div class="p-4">
                 <div>1 {{ count % 3 == 0 ? count : '' }}</div>
                 <div>2 {{ count % 3 == 1 ? count : '' }}</div>
                 <div>3 {{ count % 3 == 2 ? count : '' }}</div>
-            </div> -->
-            <!-- <div data-id="{{ count }}">Hello World!</div> -->
-            <!-- <div>
+            </div>
+            <div data-id="{{ count }}">Hello World!</div>
+            <div>
                 <input
                     class="checkbox"
                     type="checkbox"
@@ -68,7 +68,7 @@
                 <button class="btn btn-primary" onclick="toggleChecked()">
                     Toggle
                 </button>
-            </div> -->
+            </div>
             <div class="flex items-center">
                 <button onclick="increment()" class="btn btn-success">
                     Increment
@@ -82,14 +82,14 @@
                 <div class="flex-1">
                     <div class="p-2">
                         {{ new Array(count).fill(0).map( (n, i) => `
-                        <x-child2 data-count="{{ count }}" data-child="${i}"
-                            >{{ count }}</x-child2
+                        <x-child2 data-count="${count}" data-child="${i}"
+                            >${ count }</x-child2
                         >
                         `)}}
                     </div>
                 </div>
                 <div class="flex-1">
-                    <!-- <x-parent data-parent="{{ count }}"></x-parent> -->
+                    <x-parent data-parent="{{ count }}"></x-parent>
                 </div>
             </div>
         </div>
