+++ lib/cog.js
@@ -1,1 +1,1 @@
-!function(n,t){"object"==typeof exports&&"object"==typeof module?module.exports=t():"function"==typeof define&&define.amd?define([],t):"object"==typeof exports?exports.t=t():n.t=t()}(self,(()=>(()=>{"use strict";var n={d:(t,r)=>{for(var e in r)n.i(r,e)&&!n.i(t,e)&&Object.defineProperty(t,e,{enumerable:!0,get:r[e]})},i:(n,t)=>Object.prototype.hasOwnProperty.call(n,t),r:n=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(n,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(n,"u",{value:!0})}},t={};function r(n,t){var r=n(t);return Array.isArray(r)&&(r=r.join("")),r}function e(n){for(var t=n.indexOf("{{"),r=0,e=t;e<n.length;e++)if("{"===n[e]&&"{"===n[e+1]?(r++,e++):"}"===n[e]&&"}"===n[e+1]&&(r--,e++),0===r)return{start:t,end:e};return{start:t,end:-1}}n.r(t),n.d(t,{init:()=>H,variable:()=>J});var i={},u=function(n,t){try{var r=n+JSON.stringify(Object.keys(t).join(""));if(!i[r]){var e="return (state) => {".concat(Object.keys(t).map((function(n){return"const ".concat(n,' = state["').concat(n,'"];')})).join("\n")," return ").concat(n,"}");i[r]=Function(e)()}return i[r]}catch(t){throw new Error("Failed to create function from expression {{".concat(n,"}}: ").concat(t.message))}},o=function(n,t,e){for(var i=n,o="",a=0;a<t.length;a++){var f=t[a],c=f.start,l=f.end,v=f.value,s=i.slice(0,c),d=i.slice(l+1),h=r(u(v,e),e);o+="".concat(s).concat(h),i=d}return o+i},a=function(n){for(var t,r,i=[],u=String(n),o=!0;o;){var a=e(u),f=a.start,c=a.end;if(-1===c){o=!1;break}var l=u.slice(f+2,c-1),v=u.slice(c+1),s=(t=l,r=void 0,(r=document.createElement("div")).innerHTML=function(n){return n.replace(/<(?=[^<>]*>)/g,"&lt;").replace(/(?<=[^<>]*)>/g,"&gt;")}(t),(r.textContent||r.innerText||"").replace(/[\r\n]+/g,""));i.push({start:f,end:c,value:s}),u=v}return i},f=function(n){return n.split("-").reduce((function(n,t,r){return n+(r?t[0].toUpperCase()+t.slice(1):t)}),"")};function c(n){return"true"===n||"false"!==n&&("null"===n?null:"undefined"===n?void 0:""===n?"":isNaN(Number(n))?n:Number(n))}var l={};function v(n,t){var r=JSON.stringify(n)+JSON.stringify(t);if(!l[r]){for(var e=Object.assign({},t),i=0;i<n.length;i++)e[f(n[i].name)]=c(n[i].o?o(n[i].value,n[i].l,t):n[i].value);l[r]=e}return l[r]}var s=function(n){return Array.from(n.attributes).map((function(n){var t=a(n.value);return{name:n.name,value:n.value,l:t,o:!!t.length}}))},d=function(n){return n.replace(/[\r\n]+\s*/g,"")};function h(n){return(new DOMParser).parseFromString(n,"text/html").body.firstChild}function m(n,t,r,e,i){var u=arguments.length>5&&void 0!==arguments[5]?arguments[5]:[],o=arguments.length>6&&void 0!==arguments[6]?arguments[6]:null,a=arguments.length>7&&void 0!==arguments[7]?arguments[7]:[];t.add({id:n,parentId:o,element:r,v:e,lastTemplateEvaluation:i,attributes:u,l:a,shouldUpdate:!1})}function b(n,t,r){for(var e=function(n){for(var t=[],r=document.evaluate("template",n,null,XPathResult.ORDERED_NODE_ITERATOR_TYPE,null),e=r.iterateNext();e;)t.push(e),e=r.iterateNext();return t}(n),i=document.createDocumentFragment(),u=0;u<e.length;u++){var o=e[u].getAttribute("id");if(o){if(e[u].innerHTML=d(e[u].innerHTML),1!==e[u].content.childNodes.length)throw new Error("Template ".concat(o," should have a single child"));p(o,e[u],t,r),i.appendChild(e[u])}}i.textContent=""}function p(n,t,r,e){function i(){return Reflect.construct(HTMLElement,[],i)}i.prototype=Object.create(HTMLElement.prototype),i.prototype.constructor=i,i.prototype.connectedCallback=function(n,t,r){return function(){var e,i=r.id(),u=s(this),f=a(this.innerHTML);u.push({name:"children",value:this.innerHTML,l:f,o:!!f.length});var c=h(n.innerHTML);if(c.nodeType!==Node.TEXT_NODE)for(var l=c.querySelectorAll("*"),v=0;v<l.length;v++){var d=l[v];d.tagName.includes("-")&&d.setAttribute("data-parent-id",String(i))}var b=c.outerHTML;b||(b=c.textContent||"");var p=this.dataset.parentId?Number(this.dataset.parentId):null,y=h(b),x=a(b);try{var N=g(p,u,t,r.list);y=h(o(b,x,N))}catch(n){}y.cogAnchorId=i,null===(e=this.parentElement)||void 0===e||e.replaceChild(y,this),m(i,r,y,b,null,u,p,x)}}(t,r,e),customElements.define(n,i)}function g(n,t,r,e){var i=e.find((function(t){return t.id===n}));if(!i)return v(t,r);var u=g(i.parentId,i.attributes,r,e);return Object.assign({},u,v(t,u))}var y=function(n,t,r){var e=t.getAttribute("data-on-".concat(n));if(!e)throw new Error("Missing data-handler attribute");var i=u(e,r);return function(t){try{i(r),t.preventDefault()}catch(t){throw new Error("".concat(t.message,": data-on-").concat(n,"=").concat(e))}}};function x(n,t,r){n.querySelectorAll("[data-on-".concat(t,"]")).forEach((function(n){var e=y(t,n,r);n.addEventListener(t,e),n["".concat(t,"Handler")]=e}))}function N(n,t){x(n,"click",t),x(n,"change",t)}var O=function(n){return n.nodeType!==Node.TEXT_NODE&&-1!==n.tagName.indexOf("-")},j=function(n,t){for(var r=[],e=document.evaluate("self::*[text()[contains(., '{{')] and text()[contains(., '}}')]] | self::*[@*[contains(., '{{') and contains(., '}}')]] | .//*[text()[contains(., '{{')] and text()[contains(., '}}')]] | .//*[@*[contains(., '{{') and contains(., '}}')]]",n,null,XPathResult.ORDERED_NODE_ITERATOR_TYPE,null),i=e.iterateNext();i;)O(i)||r.push(i),i=e.iterateNext();for(var u=0;u<r.length;u++){var o=t.id(),f=r[u];f.innerHTML=f.innerHTML.trim();var c=f.outerHTML,l=a(c);m(o,t,f,c,-1!==c.indexOf("data-attribute")?c:null,[],null,l)}};function w(n,t){n.querySelectorAll("[data-on-".concat(t,"]")).forEach((function(n){var r=n["".concat(t,"Handler")];r&&n.removeEventListener(t,r)}))}function E(n,t){if(n.nodeType===Node.TEXT_NODE)return function(n,t){var r;return n.textContent!==t.textContent?[{node:n,content:null!==(r=t.textContent)&&void 0!==r?r:""}]:[]}(n,t);n.innerHTML=d(n.innerHTML),t.innerHTML=d(t.innerHTML);var r=function(n,t){for(var r=[],e=0;e<n.attributes.length;e++){var i=n.attributes[e],u=t.getAttribute(i.name);u!==i.value&&r.push({name:i.name,newValue:c(u||"")})}return r}(n,t);return(r.length>0?[{node:n,attributes:r}]:[]).concat(function(n,t){for(var r=[],e=!1,i=[],u=[],o=Math.max(n.childNodes.length,t.childNodes.length),a=0;a<o;a++){var f,c,l=n.childNodes[a],v=t.childNodes[a];if(void 0!==l&&l.nodeType===Node.TEXT_NODE&&void 0!==v&&v.nodeType===Node.TEXT_NODE){if((null===(f=l.textContent)||void 0===f?void 0:f.trim())!==(null===(c=v.textContent)||void 0===c?void 0:c.trim())){e=!0;break}}else void 0===l&&void 0!==v?u.push(v):void 0!==l&&void 0===v?i.push(l):r=r.concat(E(l,v))}return i.length>0&&r.push({node:n,h:i}),u.length>0&&r.push({node:n,m:u}),e?[{node:n,content:t.innerHTML}]:r}(n,t))}function S(n,t,r){for(var e=[],i=n;i!==t;)e.unshift(Array.prototype.indexOf.call(i.parentNode.childNodes,i)),i=i.parentNode;for(var u=r,o=0;o<e.length;o++){var a=e[o];if(!u.childNodes[a])return null;u=u.childNodes[a]}return u}function M(n,t){if(t.name.startsWith("data-attribute-")){var r=t.name.substring(15);t.newValue?(n[r]=!0,n.setAttribute(r,t.newValue)):(n[r]=!1,n.removeAttribute(r))}}var A=function(n,t,r,e,i,u,o,f){if(O(t)){var c,l=null!==(c=null==e?void 0:e.slice())&&void 0!==c?c:[];if(void 0!==r&&l.push({name:"children",newValue:r}),l.length){var v=function(n){return n.map((function(n){var t=a(n.newValue);return{name:n.name,value:n.newValue,l:t,o:!!t.length}}))}(l),s=f.index[n.cogAnchorId],d=f.list[s];f.update(s,"attributes",function(n,t){for(var r=n.concat(t),e={},i=0;i<r.length;i++)e[r[i].name]=r[i];return Object.values(e)}(d.attributes,v))}}else{if(void 0!==r)n.nodeType===Node.TEXT_NODE?n.textContent=r:(w(g=n,"click"),w(g,"change"),n.innerHTML=r,N(n,o));else if(void 0!==e)for(var h=0;h<e.length;h++)M(n,e[h]),n.setAttribute(e[h].name,e[h].newValue);else if(i.length){for(var m=document.createDocumentFragment(),b=0;b<i.length;b++)m.appendChild(i[b]);n.appendChild(m)}else if(u.length)for(var p=0;p<u.length;p++)n.removeChild(u[p]);var g}};function T(n,t,r,e){if(null===n)return v(t,r);var i=e.find((function(t){return t.id===n})),u=T(i.parentId,i.attributes,r,e);return Object.assign({},u,v(t,u))}function R(n,t,r){if(null===n)return t;var e=r.find((function(t){return t.id===n}));return R(e.parentId,e.attributes,r).concat(t)}var k=function(n,t){return n.some((function(n){return new RegExp("(\\s".concat(n,"\\s|{").concat(n,"\\s|\\s").concat(n,"}|{").concat(n,"}|(").concat(n,"))"),"gm").test(t)}))};function F(){return{p:0,list:[],index:{},get value(){return this.list},add:function(n){this.list.push(n),this.index[n.id]=this.list.length-1},update:function(n,t,r){"attributes"===t&&(this.list[n].shouldUpdate=!0),this.list[n][t]=r},N:function(n){this.list=n.filter((function(n){var t=n.element;return document.body.contains(t)})),this.index=this.list.reduce((function(n,t,r){return n[t.id]=r,n}),{})},id:function(){return this.p++}}}var H=function(n){var t=F(),r=function(n){return{element:null,get value(){if(this.element||(this.element=n.querySelector("#app")),!this.element)throw new Error("No app element found!");return this.element}}}(n),e=null,i={state:null,O:[],get value(){return this.state||(this.state={}),this.state},set:function(n,t){this.state||(this.state={}),this.state[n]=t,this.O.push(n)},j:function(){this.O=[]}};function u(n,r){i.set(n,r),null!==e&&cancelAnimationFrame(e),e=requestAnimationFrame((function(){!function(n,t,r){for(var e=0;e<n.value.length;e++){var i=n.value[e],u=i.id,a=i.parentId,f=i.attributes,c=i.element,l=i.v,v=i.lastTemplateEvaluation,s=i.l,d=i.shouldUpdate,m=R(a,f,n.list);if(k(r,l+" "+m.map((function(n){return n.value})).join(" "))||d){d&&n.update(e,"shouldUpdate",!1);var b=null,p=T(a,f,t,n.list);try{b=o(l,s,p)}catch(n){console.error(n);continue}var g=h(b);if(null===v){var y;n.update(e,"lastTemplateEvaluation",b),n.update(e,"element",g),g.cogAnchorId=u,null===(y=c.parentNode)||void 0===y||y.replaceChild(g,c)}else{var x=h(v),N=E(x,g);if(N.length>0){n.update(e,"lastTemplateEvaluation",b);for(var O=0;O<N.length;O++){var j=S(N[O].node,x,c),w=[],M=[];if(void 0!==N[O].m&&(M=N[O].m),void 0!==N[O].h)for(var F=0;F<N[O].h.length;F++){var H=S(N[O].h[F],x,c);H&&w.push(H)}A(j,N[O].node,N[O].content,N[O].attributes,M,w,p,n)}}}}}n.N(n.list)}(t,i.value,i.O),i.j()}))}return j(r.value,t),b(r.value,i.value,t),N(r.value,i.value),{variable:function(n,t){return u(n,t),{set value(t){u(n,t)},get value(){return i.value[n]},set:function(t){u(n,t)}}}}},J=H(document).variable;return t})()));
+!function(n,t){"object"==typeof exports&&"object"==typeof module?module.exports=t():"function"==typeof define&&define.amd?define([],t):"object"==typeof exports?exports.t=t():n.t=t()}(self,(()=>(()=>{"use strict";var n={d:(t,r)=>{for(var e in r)n.i(r,e)&&!n.i(t,e)&&Object.defineProperty(t,e,{enumerable:!0,get:r[e]})},i:(n,t)=>Object.prototype.hasOwnProperty.call(n,t),r:n=>{"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(n,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(n,"u",{value:!0})}},t={};function r(n,t){var r=n(t);return Array.isArray(r)&&(r=r.join("")),r}function e(n){for(var t=n.indexOf("{{"),r=0,e=t;e<n.length;e++)if("{"===n[e]&&"{"===n[e+1]?(r++,e++):"}"===n[e]&&"}"===n[e+1]&&(r--,e++),0===r)return{start:t,end:e};return{start:t,end:-1}}n.r(t),n.d(t,{init:()=>H,variable:()=>J});var i={},u=function(n,t){try{var r=n+JSON.stringify(Object.keys(t).join(""));if(!i[r]){var e="return (state) => {".concat(Object.keys(t).map((function(n){return"const ".concat(n,' = state["').concat(n,'"];')})).join("\n")," return ").concat(n,"}");i[r]=Function(e)()}return i[r]}catch(t){throw new Error("Failed to create function from expression {{".concat(n,"}}: ").concat(t.message))}},o=function(n,t,e){for(var i=n,o="",a=0;a<t.length;a++){var f=t[a],c=f.start,l=f.end,v=f.value,s=i.slice(0,c),d=i.slice(l+1),h=r(u(v,e),e);o+="".concat(s).concat(h),i=d}return o+i},a=function(n){for(var t,r,i=[],u=String(n),o=!0;o;){var a=e(u),f=a.start,c=a.end;if(-1===c){o=!1;break}var l=u.slice(f+2,c-1),v=u.slice(c+1),s=(t=l,r=void 0,(r=document.createElement("div")).innerHTML=function(n){return n.replace(/<(?=[^<>]*>)/g,"&lt;").replace(/(?<=[^<>]*)>/g,"&gt;")}(t),(r.textContent||r.innerText||"").replace(/[\r\n]+/g,""));i.push({start:f,end:c,value:s}),u=v}return i},f=function(n){return Array.from(n.attributes).map((function(n){var t=a(n.value);return{name:n.name,value:n.value,o:t,l:!!t.length}}))},c=function(n){return n.split("-").reduce((function(n,t,r){return n+(r?t[0].toUpperCase()+t.slice(1):t)}),"")};function l(n){return"true"===n||"false"!==n&&("null"===n?null:"undefined"===n?void 0:""===n?"":isNaN(Number(n))?n:Number(n))}var v={};function s(n,t){var r=JSON.stringify(n)+JSON.stringify(t);if(!v[r]){for(var e=Object.assign({},t),i=0;i<n.length;i++)e[c(n[i].name)]=l(n[i].l?o(n[i].value,n[i].o,t):n[i].value);v[r]=e}return v[r]}function d(n,t,r,e){var i=e.find((function(t){return t.id===n}));if(!i)return s(t,r);var u=d(i.parentId,i.attributes,r,e);return Object.assign({},u,s(t,u))}var h=function(n){return n.replace(/[\r\n]+\s*/g,"")};function m(n){return(new DOMParser).parseFromString(n,"text/html").body.firstChild}function b(n,t,r,e,i){var u=arguments.length>5&&void 0!==arguments[5]?arguments[5]:[],o=arguments.length>6&&void 0!==arguments[6]?arguments[6]:null,a=arguments.length>7&&void 0!==arguments[7]?arguments[7]:[];t.add({id:n,parentId:o,element:r,v:e,lastTemplateEvaluation:i,attributes:u,o:a,shouldUpdate:!1})}function p(n,t,r){for(var e=function(n){for(var t=[],r=document.evaluate("template",n,null,XPathResult.ORDERED_NODE_ITERATOR_TYPE,null),e=r.iterateNext();e;)t.push(e),e=r.iterateNext();return t}(n),i=document.createDocumentFragment(),u=0;u<e.length;u++){var o=e[u].getAttribute("id");if(o){if(e[u].innerHTML=h(e[u].innerHTML),1!==e[u].content.childNodes.length)throw new Error("Template ".concat(o," should have a single child"));g(o,e[u],t,r),i.appendChild(e[u])}}i.textContent=""}function g(n,t,r,e){function i(){return Reflect.construct(HTMLElement,[],i)}i.prototype=Object.create(HTMLElement.prototype),i.prototype.constructor=i,i.prototype.connectedCallback=function(n,t,r){return function(){var e,i=r.id(),u=f(this),c=a(this.innerHTML);u.push({name:"children",value:this.innerHTML,o:c,l:!!c.length});var l=m(n.innerHTML);if(l.nodeType!==Node.TEXT_NODE)for(var v=l.querySelectorAll("*"),s=0;s<v.length;s++){var h=v[s];h.tagName.includes("-")&&h.setAttribute("data-parent-id",String(i))}var p=l.outerHTML;p||(p=l.textContent||"");var g=this.dataset.parentId?Number(this.dataset.parentId):null,y=m(p),x=a(p);try{var N=d(g,u,t,r.list);y=m(o(p,x,N))}catch(n){}y.cogAnchorId=i,null===(e=this.parentElement)||void 0===e||e.replaceChild(y,this),b(i,r,y,p,null,u,g,x)}}(t,r,e),customElements.define(n,i)}var y=function(n,t,r){var e=t.getAttribute("data-on-".concat(n));if(!e)throw new Error("Missing data-handler attribute");var i=u(e,r);return function(t){try{i(r),t.preventDefault()}catch(t){throw new Error("".concat(t.message,": data-on-").concat(n,"=").concat(e))}}};function x(n,t,r){n.querySelectorAll("[data-on-".concat(t,"]")).forEach((function(n){var e=y(t,n,r);n.addEventListener(t,e),n["".concat(t,"Handler")]=e}))}function N(n,t){x(n,"click",t),x(n,"change",t)}var w=function(n){return n.nodeType!==Node.TEXT_NODE&&-1!==n.tagName.indexOf("-")},O=function(n,t){for(var r=[],e=document.evaluate("self::*[text()[contains(., '{{')] and text()[contains(., '}}')]] | self::*[@*[contains(., '{{') and contains(., '}}')]] | .//*[text()[contains(., '{{')] and text()[contains(., '}}')]] | .//*[@*[contains(., '{{') and contains(., '}}')]]",n,null,XPathResult.ORDERED_NODE_ITERATOR_TYPE,null),i=e.iterateNext();i;)w(i)||r.push(i),i=e.iterateNext();for(var u=0;u<r.length;u++){var o=t.id(),f=r[u];f.innerHTML=f.innerHTML.trim();var c=f.outerHTML,l=a(c);b(o,t,f,c,-1!==c.indexOf("data-attribute")?c:null,[],null,l)}};function j(n,t){n.querySelectorAll("[data-on-".concat(t,"]")).forEach((function(n){var r=n["".concat(t,"Handler")];r&&n.removeEventListener(t,r)}))}function E(n,t){if(n.nodeType===Node.TEXT_NODE)return function(n,t){var r;return n.textContent!==t.textContent?[{node:n,content:null!==(r=t.textContent)&&void 0!==r?r:""}]:[]}(n,t);n.innerHTML=h(n.innerHTML),t.innerHTML=h(t.innerHTML);var r=function(n,t){for(var r=[],e=0;e<n.attributes.length;e++){var i=n.attributes[e],u=t.getAttribute(i.name);u!==i.value&&r.push({name:i.name,newValue:l(u||"")})}return r}(n,t);return(r.length>0?[{node:n,attributes:r}]:[]).concat(function(n,t){for(var r=[],e=!1,i=[],u=[],o=Math.max(n.childNodes.length,t.childNodes.length),a=0;a<o;a++){var f,c,l=n.childNodes[a],v=t.childNodes[a];if(void 0!==l&&l.nodeType===Node.TEXT_NODE&&void 0!==v&&v.nodeType===Node.TEXT_NODE){if((null===(f=l.textContent)||void 0===f?void 0:f.trim())!==(null===(c=v.textContent)||void 0===c?void 0:c.trim())){e=!0;break}}else void 0===l&&void 0!==v?u.push(v):void 0!==l&&void 0===v?i.push(l):r=r.concat(E(l,v))}return i.length>0&&r.push({node:n,h:i}),u.length>0&&r.push({node:n,m:u}),e?[{node:n,content:t.innerHTML}]:r}(n,t))}function S(n,t,r){for(var e=[],i=n;i!==t;)e.unshift(Array.prototype.indexOf.call(i.parentNode.childNodes,i)),i=i.parentNode;for(var u=r,o=0;o<e.length;o++){var a=e[o];if(!u.childNodes[a])return null;u=u.childNodes[a]}return u}function M(n,t){if(t.name.startsWith("data-attribute-")){var r=t.name.substring(15);t.newValue?(n[r]=!0,n.setAttribute(r,t.newValue)):(n[r]=!1,n.removeAttribute(r))}}var A=function(n,t,r,e,i,u,o,f){if(w(t)){var c,l=null!==(c=null==e?void 0:e.slice())&&void 0!==c?c:[];if(void 0!==r&&l.push({name:"children",newValue:r}),l.length){var v=function(n){return n.map((function(n){var t=a(n.newValue);return{name:n.name,value:n.newValue,o:t,l:!!t.length}}))}(l),s=f.index[n.cogAnchorId],d=f.list[s];f.update(s,"attributes",function(n,t){for(var r=n.concat(t),e={},i=0;i<r.length;i++)e[r[i].name]=r[i];return Object.values(e)}(d.attributes,v))}}else{if(void 0!==r)n.nodeType===Node.TEXT_NODE?n.textContent=r:(j(g=n,"click"),j(g,"change"),n.innerHTML=r,N(n,o));else if(void 0!==e)for(var h=0;h<e.length;h++)M(n,e[h]),n.setAttribute(e[h].name,e[h].newValue);else if(i.length){for(var m=document.createDocumentFragment(),b=0;b<i.length;b++)m.appendChild(i[b]);n.appendChild(m)}else if(u.length)for(var p=0;p<u.length;p++)n.removeChild(u[p]);var g}};function T(n,t,r){if(null===n)return t;var e=r.find((function(t){return t.id===n}));return T(e.parentId,e.attributes,r).concat(t)}var R=function(n){return new RegExp("(\\s".concat(n,"\\s|{").concat(n,"\\s|\\s").concat(n,"}|{").concat(n,"}|(").concat(n,"))"),"gm")},k=function(n,t){return n.map(R).some((function(n){return n.test(t)}))};function F(){return{p:0,list:[],index:{},get value(){return this.list},add:function(n){this.list.push(n),this.index[n.id]=this.list.length-1},update:function(n,t,r){"attributes"===t&&(this.list[n].shouldUpdate=!0),this.list[n][t]=r},N:function(n){this.list=n.filter((function(n){var t=n.element;return document.body.contains(t)})),this.index=this.list.reduce((function(n,t,r){return n[t.id]=r,n}),{})},id:function(){return this.p++}}}var H=function(n){var t=F(),r=function(n){return{element:null,get value(){if(this.element||(this.element=n.querySelector("#app")),!this.element)throw new Error("No app element found!");return this.element}}}(n),e=null,i={state:null,O:[],get value(){return this.state||(this.state={}),this.state},set:function(n,t){this.state||(this.state={}),this.state[n]=t,this.O.push(n)},j:function(){this.O=[]}};function u(n,r){i.set(n,r),null!==e&&cancelAnimationFrame(e),e=requestAnimationFrame((function(){!function(n,t,r){for(var e=0;e<n.value.length;e++){var i=n.value[e],u=i.id,a=i.parentId,f=i.attributes,c=i.element,l=i.v,v=i.lastTemplateEvaluation,s=i.o,h=i.shouldUpdate,b=T(a,f,n.list);if(k(r,l+" "+b.map((function(n){return n.value})).join(" "))||h){h&&n.update(e,"shouldUpdate",!1);var p=null,g=d(a,f,t,n.list);try{p=o(l,s,g)}catch(n){console.error(n);continue}var y=m(p);if(null===v){var x;n.update(e,"lastTemplateEvaluation",p),n.update(e,"element",y),y.cogAnchorId=u,null===(x=c.parentNode)||void 0===x||x.replaceChild(y,c)}else{var N=m(v),w=E(N,y);if(w.length>0){n.update(e,"lastTemplateEvaluation",p);for(var O=0;O<w.length;O++){var j=S(w[O].node,N,c),M=[],R=[];if(void 0!==w[O].m&&(R=w[O].m),void 0!==w[O].h)for(var F=0;F<w[O].h.length;F++){var H=S(w[O].h[F],N,c);H&&M.push(H)}A(j,w[O].node,w[O].content,w[O].attributes,R,M,g,n)}}}}}n.N(n.list)}(t,i.value,i.O),i.j()}))}return O(r.value,t),p(r.value,i.value,t),N(r.value,i.value),{variable:function(n,t){return u(n,t),{set value(t){u(n,t)},get value(){return i.value[n]},set:function(t){u(n,t)}}}}},J=H(document).variable;return t})()));
